requirements.txt
opencv-python==4.2.0.34
pyyaml
schedule